Move formatLinkLabel utility into its own class, with a test
